chore: Update image and date props in Home.jsx
This commit updates the image and date props in the Home.jsx file to match the changes made in the data structure. The image prop has been changed to blog?.mainPhoto and the date prop has been changed to blog?.publicationDate. These updates ensure that the component receives the correct data and improves the accuracy of the displayed information.